% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{.scale_oucru}
\alias{.scale_oucru}
\title{Internal constructor for discrete or continuous ggplot scales}
\usage{
.scale_oucru(aes, type = c("d", "c"), palette = "main", direction = 1, ...)
}
\arguments{
\item{aes}{The ggplot aesthetic mapping, e.g. "fill" or "color"}

\item{type}{Aesthetic scale type, either "d" (discrete) or "c" (continuous)}

\item{palette}{OUCRU palette name}

\item{direction}{Palette direction, 0 or positive number means forward, negative means backward}

\item{...}{Passed to \code{ggplot2::discrete_scale()} or \code{ggplot2::continuous_scale()}}
}
\value{
\code{ggplot2::discrete_scale()} or \code{ggplot2::continuous_scale()}
}
\description{
Internal constructor for discrete or continuous ggplot scales
}
\keyword{internal}
