% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{ggplot_scales}
\alias{ggplot_scales}
\alias{scale_fill_oucru_c}
\alias{scale_color_oucru_c}
\alias{scale_colour_oucru_c}
\alias{scale_fill_oucru_d}
\alias{scale_color_oucru_d}
\alias{scale_colour_oucru_d}
\title{OUCRU colors for \code{ggplot2} scales}
\usage{
scale_fill_oucru_c(palette = "main", direction = 1, ...)

scale_color_oucru_c(palette = "main", direction = 1, ...)

scale_colour_oucru_c(palette = "main", direction = 1, ...)

scale_fill_oucru_d(palette = "main", direction = 1, ...)

scale_color_oucru_d(palette = "main", direction = 1, ...)

scale_colour_oucru_d(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{OUCRU palette name}

\item{direction}{Palette direction, 0 or positive number means forward, negative means backward}

\item{...}{Passed to \code{ggplot2::continuous_scale()}}
}
\description{
OUCRU colors for \code{ggplot2} scales
}
