% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcmc_shapefiles.R
\docType{data}
\name{hcmc_shapefiles}
\alias{hcmc_shapefiles}
\title{Ho Chi Minh City shapefiles data}
\format{
\subsection{\code{hcmc_shapefiles}}{

A named list with 5 objects
\describe{
\item{district_24}{24 districts in HCM city (during 2003 - 2021)}
\item{commune_322}{322 communes in HCM city (during 2003 - 2021)}
\item{commune_312}{312 communes in HCM city (during 2021 - 2023)}
\item{commune_168}{168 communes in HCM city (after 2025 reform)}
\item{commune_167_mainland}{Same as \code{commune_168} but only mainland
communes (no Con Dao island)}
}
}
}
\source{
Derived from \url{https://quingzz.github.io/hcmc_geo_change/}
}
\usage{
hcmc_shapefiles
}
\description{
A list of shapefiles (as \code{sf} objects) for Ho Chi Minh City
}
\keyword{datasets}
