% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oucru-package.R
\docType{package}
\name{oucru-package}
\alias{oucru}
\alias{oucru-package}
\title{oucru: Collection of helper and miscellaneous functions used for and by the Oxford University Clinical Research Unit (OUCRU), Vietnam}
\description{
Collection of helper and miscellaneous functions used for and by the Oxford University Clinical Research Unit (OUCRU), Vietnam.
}
\author{
\strong{Maintainer}: Tuyen Huynh \email{tuyenhn@oucru.org}

}
\keyword{internal}
