% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_and_pal.R
\name{oucru_color}
\alias{oucru_color}
\title{Fetch OUCRU color by names (if it is defined)}
\usage{
oucru_color(...)
}
\arguments{
\item{...}{Color names as strings}
}
\value{
Named list of colors, or full color list if \code{...} is NULL
}
\description{
Fetch OUCRU color by names (if it is defined)
}
\examples{
# example code
oucru_color("blue", "orange")
}
