% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_and_pal.R
\docType{data}
\name{OUCRU_COLS}
\alias{OUCRU_COLS}
\title{Define OUCRU colors based on the brand guideline}
\format{
An object of class \code{character} of length 18.
}
\usage{
OUCRU_COLS
}
\description{
Define OUCRU colors based on the brand guideline
}
\keyword{datasets}
