% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_and_pal.R
\docType{data}
\name{oucru_cols_and_pal}
\alias{oucru_cols_and_pal}
\alias{OUCRU_COLS}
\alias{oucru_color}
\alias{OUCRU_PALS}
\alias{oucru_palette}
\alias{oucru_list_pals}
\title{Functions and constants defining OUCRU colors and palettes based on the
brand guideline}
\format{
An object of class \code{character} of length 18.

An object of class \code{list} of length 7.
}
\usage{
OUCRU_COLS

oucru_color(...)

OUCRU_PALS

oucru_palette(palette = "main")

oucru_list_pals()
}
\arguments{
\item{...}{Color names as strings}

\item{palette}{Palette name, check available palettes with \code{oucru_list_pals()}}
}
\value{
Named list of colors, or full color list if \code{...} is NULL

OUCRU color palette
}
\description{
Functions and constants defining OUCRU colors and palettes based on the
brand guideline
}
\examples{
# example code
oucru::oucru_color("blue", "orange")

# example code
oucru::oucru_palette("main")

}
\keyword{datasets}
