% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_and_pal.R
\name{oucru_palette}
\alias{oucru_palette}
\title{Fetch OUCRU color palette by name (if it is defined)}
\usage{
oucru_palette(palette = "main", ...)
}
\arguments{
\item{palette}{Palette name}

\item{...}{unused}
}
\value{
OUCRU color palette
}
\description{
Fetch OUCRU color palette by name (if it is defined)
}
\examples{
# example code
oucru::oucru_palette("main")
}
