% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_and_pal.R
\docType{data}
\name{OUCRU_PALS}
\alias{OUCRU_PALS}
\title{Define OUCRU palettes based on the brand guideline}
\format{
An object of class \code{list} of length 7.
}
\usage{
OUCRU_PALS
}
\description{
Define OUCRU palettes based on the brand guideline
}
\keyword{datasets}
