% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_and_pal.R
\name{palette_gen_c}
\alias{palette_gen_c}
\title{Palette generator for continuous scale (used by \code{ggplot2})}
\usage{
palette_gen_c(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{OUCRU palette name}

\item{direction}{Palette direction, 0 or positive number means forward, negative means backward}

\item{...}{Passed to \code{grDevices::colorRampPalette()}}
}
\description{
Palette generator for continuous scale (used by \code{ggplot2})
}
