% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_and_pal.R
\name{palette_gen_d}
\alias{palette_gen_d}
\title{Palette generator for use by \code{ggplot2}}
\usage{
palette_gen_d(palette = "main", direction = 1)
}
\arguments{
\item{palette}{OUCRU palette name}

\item{direction}{Palette direction, 0 or positive number means forward, negative means backward}
}
\description{
Palette generator for use by \code{ggplot2}
}
\keyword{internal}
