% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales_c.R
\name{scale_color_oucru_c}
\alias{scale_color_oucru_c}
\title{\code{ggplot2} geom colors for continuous scale}
\usage{
scale_color_oucru_c(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{OUCRU palette name}

\item{direction}{Palette direction, 0 or positive number means forward, negative means backward}
}
\description{
\code{ggplot2} geom colors for continuous scale
}
