% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scales.R
\name{scale_fill_oucru_d}
\alias{scale_fill_oucru_d}
\title{\code{ggplot2} fill colors for discrete scale}
\usage{
scale_fill_oucru_d(palette = "main", direction = 1, ...)
}
\arguments{
\item{palette}{OUCRU palette name}

\item{direction}{Palette direction, 0 or positive number means forward, negative means backward}

\item{...}{Passed to \code{ggplot2::continuous_scale()}}
}
\description{
\code{ggplot2} fill colors for discrete scale
}
