% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_and_pal.R
\name{oucru_cols_and_pal}
\alias{oucru_cols_and_pal}
\alias{oucru_color}
\alias{oucru_palette}
\alias{oucru_list_pals}
\title{Functions to fetch OUCRU colors and palettes based on the brand guideline}
\usage{
oucru_color(...)

oucru_palette(palette = "main")

oucru_list_pals()
}
\arguments{
\item{...}{Color names as strings}

\item{palette}{Palette name, check available palettes with \code{oucru_list_pals()}}
}
\description{
Functions to fetch OUCRU colors and palettes based on the brand guideline
}
\examples{
oucru::oucru_color("blue", "orange")

oucru::oucru_palette("main")

oucru::oucru_list_pals()

}
